Fixed crashing when primary lang script not found; UVrotate speed is now read as a signed byte; fixed mixup with numTotalLevels and numLevelPaths

+++ src/Gameflow.h
@@ -32,7 +32,8 @@ namespace TRLEScript
 
 	private:
 		uint16_t GetPathStringOffset(int lastIndex);
-		uint8_t GetNumUniqueLevelPaths() const;
+		uint16_t GetNumUniqueLevelPaths() const;
+		//uint8_t GetNumUniqueLevelPaths() const;
 		int GetPathIndex(const char *path);// const;
 		bool created;
